refactor: Move around profile import/export logic
