--- src/lib/impamp2Types.ts
@@ -1,46 +1,0 @@
-/**
- * Defines TypeScript interfaces representing the structure of the legacy
- * impamp2 JSON export format.
- */
-
-/**
- * Represents a single pad within an impamp2 page.
- * Note: The 'file' property contains a data URL string.
- */
-export interface Impamp2Pad {
-  page: string; // Page number as a string (e.g., "0", "1")
-  key: string; // Keyboard key associated with the pad (e.g., "q", "a", ";")
-  name: string; // Display name of the pad/sound
-  file: string; // Data URL string (e.g., "data:audio/mpeg;base64,<BASE_64_STRING>")
-  filename: string; // Original filename
-  filesize: number; // File size in bytes
-  startTime: number | null; // Start time for playback (likely unused in import)
-  endTime: number | null; // End time for playback (likely unused in import)
-  updatedAt: number; // Timestamp of last update
-  readable: boolean; // Indicates if the file is readable (likely always true for export)
-}
-
-/**
- * Represents a single page (bank) within an impamp2 export.
- * Pads are stored in an object keyed by the keyboard character.
- */
-export interface Impamp2Page {
-  pageNo: string; // Page number as a string (e.g., "0", "1")
-  name: string; // Name of the page/bank
-  emergencies: number; // Count of emergencies (likely unused in import)
-  updatedAt: number; // Timestamp of last update
-  pads: {
-    [key: string]: Impamp2Pad; // Pads keyed by keyboard character (e.g., "'", ";", "a", "b")
-  };
-}
-
-/**
- * Represents the top-level structure of an impamp2 export file.
- * Pages are stored in an object keyed by the page number string.
- */
-export interface Impamp2Export {
-  padCount: number; // Total count of pads across all pages
-  pages: {
-    [pageNo: string]: Impamp2Page; // Pages keyed by page number string (e.g., "0", "1")
-  };
-}
